requirements.txt
httpx>=0.27.0
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
langgraph-checkpoint-postgres>=2.0.0
langgraph-checkpoint-sqlite>=2.0.0
psycopg[binary,pool]>=3.1.0
